+++ src/components/ColorPicker/dataHooks.ts
@@ -1,2 +1,3 @@
 export const colorPickerDataHook = 'ColorPicker';
 export const colorPickerItemDataHook = 'ColorPickerItem';
+export const colorPickerItemTooltipDataHook = 'color-picker-item-tooltip';