+++ src/components/Event/Event.visual.tsx
@@ -13,5 +13,7 @@ visualize('Event', () => {
     snap('timeShown', <Event showTime {...defaultProps} />);
     snap('fullday', <Event fullday {...defaultProps} />);
     snap('selected', <Event selected {...defaultProps} />);
+    snap('clickable', <Event onClick={() => {}} {...defaultProps} />);
+    snap('disabled', <Event onClick={() => {}} disabled {...defaultProps} />);
   });
 });