+++ src/components/Event/docs/index.story.tsx
@@ -53,6 +53,9 @@ export default {
           ...[{ title: 'More Events', source: examples.moreEventsSection }].map(
             code,
           ),
+          ...[{ title: 'Clickable', source: examples.clickableSection }].map(
+            code,
+          ),
         ],
       }),
 
@@ -98,8 +101,13 @@ export default {
                 ],
                 fonts: [
                   {
-                    label: 'Font Style',
-                    wixParam: 'textFont',
+                    label: 'Title Font',
+                    wixParam: 'titleFont',
+                    defaultFont: 'arial',
+                  },
+                  {
+                    label: 'Time Font',
+                    wixParam: 'timeFont',
                     defaultFont: 'arial',
                   },
                 ],